Add smooth (antialiased) fonts
