+++ TFT_Drivers/S6D02A1_Rotation.h
@@ -7,22 +7,22 @@
   switch (rotation) {
     case 0:
       writedata(TFT_MAD_MX | TFT_MAD_MY | TFT_MAD_BGR);
-      _width  = _width_orig;
-      _height = _height_orig;
+      _width  = _init_width;
+      _height = _init_height;
       break;
     case 1:
       writedata(TFT_MAD_MV | TFT_MAD_MY | TFT_MAD_BGR);
-      _width  = _height_orig;
-      _height = _width_orig;
+      _width  = _init_height;
+      _height = _init_width;
       break;
     case 2:
       writedata(TFT_MAD_BGR);
-      _width  = _width_orig;
-      _height = _height_orig;
+      _width  = _init_width;
+      _height = _init_height;
       break;
     case 3:
       writedata(TFT_MAD_MX | TFT_MAD_MV | TFT_MAD_BGR);
-      _width  = _height_orig;
-      _height = _width_orig;
+      _width  = _init_height;
+      _height = _init_width;
       break;
   }
