+++ TFT_eSPI.h
@@ -1,9 +1,9 @@
 /***************************************************
-  Arduino TFT graphics library targetted at ESP8266
+  Arduino TFT graphics library targeted at ESP8266
   and ESP32 based boards.
 
   This is a standalone library that contains the
-  hardware driver, the graphics funtions and the
+  hardware driver, the graphics functions and the
   proportional fonts.
 
   The larger fonts are Run Length Encoded to reduce
@@ -73,6 +73,17 @@
 #include <pgmspace.h>
 
 #include <SPI.h>
+
+#ifdef SMOOTH_FONT
+  // Call up the SPIFFS FLASH filing system for the anti-aliased fonts
+  #define FS_NO_GLOBALS
+  #include <FS.h>
+
+  #ifdef ESP32
+    #include "SPIFFS.h"
+  #endif
+#endif
+
 
 #if defined (ESP8266) && defined (D0_USED_FOR_DC)
   #define DC_C digitalWrite(TFT_DC, LOW)
@@ -255,7 +266,11 @@
 template <typename T> static inline void
 swap_coord(T& a, T& b) { T t = a; a = b; b = t; }
 
-// This is a structure to conveniently hold infomation on the default fonts
+#ifndef min
+#define min(a,b) (((a) < (b)) ? (a) : (b))
+#endif
+
+// This is a structure to conveniently hold information on the default fonts
 // Stores pointer to font character image address table, width table and height
 
 // Create a null set in case some fonts not used (to prevent crash)
@@ -318,7 +333,6 @@ const PROGMEM fontinfo fontdata [] = {
 };
 
 
-
 // Class functions and variables
 class TFT_eSPI : public Print {
 
@@ -343,9 +357,8 @@ class TFT_eSPI : public Print {
   void     setWindow(int16_t x0, int16_t y0, int16_t x1, int16_t y1),
            pushColor(uint16_t color),
            pushColor(uint16_t color, uint16_t len),
-           //pushColors(uint16_t  *data, uint8_t len),
-           pushColors(uint8_t  *data, uint32_t len),
            pushColors(uint16_t  *data, uint32_t len, bool swap = true), // With byte swap option
+           pushColors(uint8_t  *data, uint32_t len),
 
            fillScreen(uint32_t color);
 
@@ -375,7 +388,7 @@ class TFT_eSPI : public Print {
            setTextColor(uint16_t fgcolor, uint16_t bgcolor),
            setTextSize(uint8_t size),
 
-           setTextWrap(boolean wrap),
+           setTextWrap(boolean wrapX, boolean wrapY = false),
            setTextDatum(uint8_t datum),
            setTextPadding(uint16_t x_width),
 
@@ -427,13 +440,14 @@ class TFT_eSPI : public Print {
 
   uint8_t  getRotation(void),
            getTextDatum(void),
-           color332(uint16_t color565); // Convert 16 bit colour to 8 bits
+           color16to8(uint16_t color565); // Convert 16 bit colour to 8 bits
 
   int16_t  getCursorX(void),
            getCursorY(void);
 
   uint16_t fontsLoaded(void),
-           color565(uint8_t r, uint8_t g, uint8_t b);
+           color565(uint8_t r, uint8_t g, uint8_t b),
+           color8to16(uint8_t color332);  // Convert 8 bit colour to 16 bits
         
   int16_t  drawNumber(long long_num,int poX, int poY, int font),
            drawNumber(long long_num,int poX, int poY),
@@ -462,49 +476,39 @@ class TFT_eSPI : public Print {
 
   void     setAddrWindow(int32_t xs, int32_t ys, int32_t xe, int32_t ye);
 
-           // These are associated with the Touch Screen handlers
-  uint8_t  getTouchRaw(uint16_t *x, uint16_t *y);
-  uint16_t getTouchRawZ(void);
-  uint8_t  getTouch(uint16_t *x, uint16_t *y, uint16_t threshold = 600);
-
-  void     calibrateTouch(uint16_t *data, uint32_t color_fg, uint32_t color_bg, uint8_t size);
-  void     setTouch(uint16_t *data);
 
   size_t   write(uint8_t);
+
+  int32_t  cursor_x, cursor_y;
+  uint32_t textcolor, textbgcolor;
+
 
  private:
 
   inline void spi_begin() __attribute__((always_inline));
   inline void spi_end()   __attribute__((always_inline));
-  inline void spi_begin_touch() __attribute__((always_inline));
-  inline void spi_end_touch()   __attribute__((always_inline));
 
   void     readAddrWindow(int32_t xs, int32_t ys, int32_t xe, int32_t ye);
-    
+
   uint8_t  tabcolor,
            colstart = 0, rowstart = 0; // some ST7735 displays need this changed
 
-  volatile uint32_t *dcport, *csport;//, *mosiport, *clkport, *rsport;
+  volatile uint32_t *dcport, *csport;
 
-  uint32_t cspinmask, dcpinmask, wrpinmask;//, mosipinmask, clkpinmask;
+  uint32_t cspinmask, dcpinmask, wrpinmask;
 
   uint32_t lastColor = 0xFFFF;
 
-           // These are associated with the Touch Screen handlers
-  uint8_t  validTouch(uint16_t *x, uint16_t *y, uint16_t threshold = 600);
-           // Initialise with example calibration values so processor does not crash if setTouch() not called in setup()
-  uint16_t touchCalibration_x0 = 300, touchCalibration_x1 = 3600, touchCalibration_y0 = 300, touchCalibration_y1 = 3600;
-  uint8_t  touchCalibration_rotate = 1, touchCalibration_invert_x = 2, touchCalibration_invert_y = 0;
-  uint32_t _pressTime;
-  uint16_t _pressX, _pressY;
 
  protected:
 
-  int32_t  cursor_x, cursor_y, win_xe, win_ye, padX;
+  int32_t  win_xe, win_ye, padX;
 
-  uint32_t _width_orig, _height_orig; // Display w/h as input, used by setRotation()
+  uint32_t _init_width, _init_height; // Display w/h as input, used by setRotation()
   uint32_t _width, _height; // Display w/h as modified by current rotation
-  uint32_t textcolor, textbgcolor, fontsloaded, addr_row, addr_col;
+  uint32_t addr_row, addr_col;
+
+  uint32_t fontsloaded;
 
   uint8_t  glyph_ab,  // glyph height above baseline
            glyph_bb,  // glyph height below baseline
@@ -513,164 +517,32 @@ class TFT_eSPI : public Print {
            textdatum, // Text reference datum
            rotation;  // Display rotation (0-3)
 
-  bool     textwrap;   // If set, 'wrap' text at right edge of display
+  bool     textwrapX, textwrapY;   // If set, 'wrap' text at right and optionally bottom edge of display
   bool     _swapBytes; // Swap the byte order for TFT pushImage()
   bool     locked, inTransaction; // Transaction and mutex lock flags for ESP32
-  bool     _booted;
 
   int32_t  _lastColor;
 
 #ifdef LOAD_GFXFF
-  GFXfont
-    *gfxFont;
+  GFXfont  *gfxFont;
 #endif
 
-};
+// Load the Touch extension
+#ifdef TOUCH_CS
+  #include "Extensions/Touch.h"
+#endif
 
-/***************************************************************************************
-// The following button class has been ported over from the Adafruit_GFX library so
-// should be compatible.
-// A slightly different implementation in this TFT_eSPI library allows the button
-// legends to be in any font
-***************************************************************************************/
+// Load the Anti-aliased font extension
+#ifdef SMOOTH_FONT
+  #include "Extensions/Smooth_font.h"
+#endif
 
-class TFT_eSPI_Button {
+}; // End of class TFT_eSPI
 
- public:
-  TFT_eSPI_Button(void);
-  // "Classic" initButton() uses center & size
-  void     initButton(TFT_eSPI *gfx, int16_t x, int16_t y,
-  uint16_t w, uint16_t h, uint16_t outline, uint16_t fill,
-  uint16_t textcolor, char *label, uint8_t textsize);
+// Load the Button Class
+#include "Extensions/Button.h"
 
-  // New/alt initButton() uses upper-left corner & size
-  void     initButtonUL(TFT_eSPI *gfx, int16_t x1, int16_t y1,
-  uint16_t w, uint16_t h, uint16_t outline, uint16_t fill,
-  uint16_t textcolor, char *label, uint8_t textsize);
-  void     drawButton(boolean inverted = false);
-  boolean  contains(int16_t x, int16_t y);
-
-  void     press(boolean p);
-  boolean  isPressed();
-  boolean  justPressed();
-  boolean  justReleased();
-
- private:
-  TFT_eSPI *_gfx;
-  int16_t  _x1, _y1; // Coordinates of top-left corner
-  uint16_t _w, _h;
-  uint8_t  _textsize;
-  uint16_t _outlinecolor, _fillcolor, _textcolor;
-  char     _label[10];
-
-  boolean  currstate, laststate;
-};
-
-
-/***************************************************************************************
-// The following class creates Sprites in RAM, graphics can then be drawn in the Sprite
-// and rendered quickly onto the TFT screen. The class inherits the graphics functions
-// from the TFT_eSPI class. Some functions are overridden by this class so that the
-// graphics are written to the Sprite rather than the TFT.
-***************************************************************************************/
-
-class TFT_eSprite : public TFT_eSPI {
-
- public:
-
-  TFT_eSprite(TFT_eSPI *tft);
-
-           // Create a sprite of width x height pixels, return a pointer to the RAM area
-           // Sketch can cast returned value to (uint16_t*) for 16 bit depth if needed
-           // RAM required is 1 byte per pixel for 8 bit colour depth, 2 bytes for 16 bit
-  uint8_t* createSprite(int16_t width, int16_t height);  
-
-           // Delete the sprite to free up the RAM
-  void     deleteSprite(void);
-
-           // Set the colour depth to 8 or 16 bits. Can be used to change depth an existing
-           // sprite, but clears it to black, returns a new pointer if sprite is re-created.
-  uint8_t* setColorDepth(int8_t b);
-
-  void     drawPixel(uint32_t x, uint32_t y, uint32_t color);
-
-  void     drawChar(int32_t x, int32_t y, unsigned char c, uint32_t color, uint32_t bg, uint8_t size),
-
-           fillSprite(uint32_t color),
-
-           // Define a window to push 16 bit colour pixels into is a raster order
-           // Colours are converted to 8 bit if depth is set to 8
-           setWindow(int32_t x0, int32_t y0, int32_t x1, int32_t y1),
-           pushColor(uint32_t color),
-           pushColor(uint32_t color, uint16_t len),
-           // Push a pixel preformatted as a 8 or 16 bit colour (avoids conversion overhead)
-           writeColor(uint16_t color),
-
-           // Set the scroll zone, top left corner at x,y with defined width and height
-           // The colour (optional, black is default) is used to fill the gap after the scroll
-           setScrollRect(int32_t x, int32_t y, uint32_t w, uint32_t h, uint16_t color = TFT_BLACK),
-           // Scroll the defined zone dx,dy pixels. Negative values left,up, positive right,down
-           // dy is optional (default is then no up/down scroll).
-           // The sprite coordinate frame does not move because pixels are moved
-           scroll(int16_t dx, int16_t dy = 0),
-
-           drawLine(int32_t x0, int32_t y0, int32_t x1, int32_t y1, uint32_t color),
-           drawFastVLine(int32_t x, int32_t y, int32_t h, uint32_t color),
-           drawFastHLine(int32_t x, int32_t y, int32_t w, uint32_t color),
-
-           fillRect(int32_t x, int32_t y, int32_t w, int32_t h, uint32_t color),
-
-           // Set the sprite text cursor position for print class (does not change the TFT screen cursor)
-           setCursor(int16_t x, int16_t y);
-
-           // Read the colour of a pixel at x,y and return value in 565 format 
-  uint16_t readPixel(int32_t x0, int32_t y0);
-
-           // Write an image (colour bitmap) to the sprite
-  void     pushImage(int32_t x0, int32_t y0, uint32_t w, uint32_t h, uint16_t *data);
-  void     pushImage(int32_t x0, int32_t y0, uint32_t w, uint32_t h, const uint16_t *data);
-
-           // Swap the byte order for pushImage() - corrects different image endianness
-  void     setSwapBytes(bool swap);
-  bool     getSwapBytes(void);
-
-           // Push the sprite to the TFT screen, this fn calls pushImage() in the TFT class.
-           // Optionally a "transparent" colour can be defined, pixels of that colour will not be rendered
-  void     pushSprite(int32_t x, int32_t y);
-  void     pushSprite(int32_t x, int32_t y, uint16_t transparent);
-
-  int16_t  drawChar(unsigned int uniCode, int x, int y, int font),
-           drawChar(unsigned int uniCode, int x, int y);
-
-           // Return the width and height of the sprite
-  int16_t  width(void),
-           height(void);
-
-           // Used by print class to print text to cursor position
-  size_t   write(uint8_t);
-
- private:
-
-  TFT_eSPI *_tft;
-
- protected:
- 
-  uint16_t *_img;  // pointer to 16 bit sprite
-  uint8_t  *_img8; // pointer to  8 bit sprite
-  bool     _created, _bpp16; // created and bits per pixel depth flags
-
-  int32_t  _icursor_x, _icursor_y;
-  int32_t  _xs, _ys, _xe, _ye, _xptr, _yptr; // for setWindow
-  int32_t  _sx, _sy; // x,y for scroll zone
-  uint32_t _sw, _sh; // w,h for scroll zone
-  uint32_t _scolor;  // gap fill colour for scroll zone
-
-  boolean  _iswapBytes; // Swap the byte order for Sprite pushImage()
-
-  int32_t  _iwidth, _iheight; // Sprite image width and height
-
-};
-
-
+// Load the Sprite Class
+#include "Extensions/Sprite.h"
 
 #endif
