+++ User_Setup_Select.h
@@ -9,7 +9,7 @@
 // changes being needed. It also improves the portability of users sketches to other
 // hardware configurations and compatible libraries.
 //
-// Create a shortcut to this file on your desktop to permit quick access for editting.
+// Create a shortcut to this file on your desktop to permit quick access for editing.
 // Re-compile and upload after making and saving any changes to this file.
 
 // Customised User_Setup files are stored in the "User_Setups" folder.
@@ -33,6 +33,8 @@
 //#include <User_Setups/Setup10_RPi_touch_ILI9486.h>  // Setup file configured for my stock RPi TFT with touch
 //#include <User_Setups/Setup11_RPi_touch_ILI9486.h>  // Setup file configured for my stock RPi TFT with touch
 //#include <User_Setups/Setup12_M5Stack.h>           // Setup file for the ESP32 based M5Stack
+
+//#include <User_Setups/Setup99.h>
 
 //#include <User_Setups/SetupX_Template.h> // Setup file template for copying/editting
