+++ library.json
@@ -1,6 +1,6 @@
 {
   "name": "TFT_eSPI",
-  "version": "0.18.17",
+  "version": "0.18.20",
   "keywords": "tft, display, ESP8266, NodeMCU, ESP32, M5Stack, ILI9341, ST7735, ILI9163, S6D02A1, ILI9486",
   "description": "A TFT SPI graphics library for ESP8266 and ESP32",
   "repository":
