+++ library.properties
@@ -1,5 +1,5 @@
 name=TFT_eSPI
-version=0.18.17
+version=0.18.20
 author=Bodmer
 maintainer=Bodmer
 sentence=A fast TFT library for ESP8266 processors and the Arduino IDE
